requirements.txt
feedparser
PyYAML
requests
